Refactor mapping fields to new inheritance module
