+++ google_shop/views/product_mapping_view_inheritance.xml
@@ -1,0 +1,29 @@
+<?xml version="1.0" encoding="utf-8"?>
+<odoo>
+    <record id="product_mapping_form_inherit_links_images" model="ir.ui.view">
+        <field name="name">Product Mapping Form Inherit Links and Images</field>
+        <field name="model">product.mapping</field>
+        <field name="inherit_id" ref="google_shop.product_mapping_form"/>
+        <field name="arch" type="xml">
+            <xpath expr="//group/group[2]" position="inside">
+                <label class="o_form_label" for="product_shop_link" style="font-weight:bold;">Google Product Link</label>
+                <span class="pt-2">
+                    <field name="product_shop_link" readonly="1" nolabel="1"/>
+                    <button name="action_open_product_shop_link" type="object" string="Open Link" class="btn-primary" attrs="{'invisible': [('product_shop_link','=',False)]}"/>
+                </span>
+            </xpath>
+            <xpath expr="//notebook" position="inside">
+                <page string="Description">
+                    <group>
+                        <field name="google_description" readonly="True" nolabel="1"/>
+                    </group>
+                </page>
+                <page string="Additional Images">
+                    <group>
+                        <field name="additional_images" readonly="True" nolabel="1"/>
+                    </group>
+                </page>
+            </xpath>
+        </field>
+    </record>
+</odoo>
